Add data processing function and input box
Added data processing function that converts the json into an array of
desired data for displaying. Added an input box for inputting the
location of desired weather forecast. Set up empty display function for
adding date to the webpage. Will now add to the page, style, and add
an appropirate gif and loading page.